Improves enemy AI and level design
Improves the enemy AI by ensuring the Froglot enemy only targets the Player.

Adds a new level with tiles and a background.

Adds a delete save button to delete save files.

Fixes a bug where the player could fall off the map by teleporting the player to the last checkpoint when they exit the screen.